[bugfix] Three small bugfixes.
(1) Set the correct arity in the BitdataLayout constructor. (Need to eliminate bogus kind and arity fields!)

(2) Use arity of layouts rather than cfuns to determine how may different selector blocks are needed for a bitdata constructor.

(3) Include enumeration types as candidates for bitdata representation by using the components of the specMap that map already specialized versions of types to themselves instead of looking for mappings from old instances to specialized versions. This works because enumerations are treated as already specialized.