Address a bug in the calculation of alignments and bitsize values in recursive type definitions like the following:
struct  S [ status :: Stored Status | badge :: Stored Word ] aligned 4
bitdata Status/32 = Absent [0] | Present [ link :: Ref S | 1 ]

For the time being, we now require an alignment declaration on the structure definition so that we can compute the bitsize for Ref S in the (mutually recursive) definition for Status.

Also includes adjustments to wording of error messages, and to force an early abort during the initial process of calculating sizes if an error is detected to avoid running further with erroneous inputs that can cause a more severe failure.